Add files via upload
Running this script usually involves downloading python, installing libraries, then running it in the command prompt or some other IDE. Not exactly easy for anyone who is not used to programming. This batch file will both install libraries and run the programme with only a double click. And a shortcut can also be set up to allow for convenient access from desktop